Refactor complex entitlements so that long running tests get their own file to bring all CI times for entitlement test files down to about 2 minutes or less.
